// PlaybookAgent/src/generated/services/CustomConnector_AzureOpenAIService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { IOperationResult } from '@microsoft/power-apps/data';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getClient } from '@microsoft/power-apps/data';

export class CustomConnector_AzureOpenAIService {
  private static readonly dataSourceName = 'cr_5fcustomconnector_5fazureopenai_5fe435f80a6d4c3504';

  private static readonly client = getClient(dataSourcesInfo);

  /**
   * chat-completion
   */
  public static async chat_completion(api_version?: string, body?: Record<string, unknown>): Promise<IOperationResult<void>> {
    const params: { api_version?: string, body?: Record<string, unknown> } = { api_version, body };
    const result = await CustomConnector_AzureOpenAIService.client.executeAsync<{ api_version?: string, body?: Record<string, unknown> }, void>(
      {
        connectorOperation: {
          tableName: CustomConnector_AzureOpenAIService.dataSourceName,
          operationName: 'chat-completion',
          parameters: params
        },
      });
    return result;
  }
}