// PlaybookAgent/src/generated/services/CustCon_AzureDocIntService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { IOperationResult } from '@microsoft/power-apps/data';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getClient } from '@microsoft/power-apps/data';

export class CustCon_AzureDocIntService {
  private static readonly dataSourceName = 'cr_5fcustcon_5fazuredocint_5fe435f80a6d4c3504';

  private static readonly client = getClient(dataSourcesInfo);

  /**
   * Analyze document
   */
  public static async AnalyzeDocument(api_version: string, body: Record<string, unknown>, outputContentFormat?: string, features?: string): Promise<IOperationResult<void>> {
    const params: { api_version: string, body: Record<string, unknown>, outputContentFormat?: string, features?: string } = { api_version, body, outputContentFormat, features };
    const result = await CustCon_AzureDocIntService.client.executeAsync<{ api_version: string, body: Record<string, unknown>, outputContentFormat?: string, features?: string }, void>(
      {
        connectorOperation: {
          tableName: CustCon_AzureDocIntService.dataSourceName,
          operationName: 'AnalyzeDocument',
          parameters: params
        },
      });
    return result;
  }

  /**
   * Get analyze result
   */
  public static async GetAnalyzeResult(api_version: string, resultId: string): Promise<IOperationResult<void>> {
    const params: { api_version: string, resultId: string } = { api_version, resultId };
    const result = await CustCon_AzureDocIntService.client.executeAsync<{ api_version: string, resultId: string }, void>(
      {
        connectorOperation: {
          tableName: CustCon_AzureDocIntService.dataSourceName,
          operationName: 'GetAnalyzeResult',
          parameters: params
        },
      });
    return result;
  }
}