// PlaybookAgent/src/generated/services/CustCon_SAP_OdataService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { IOperationResult } from '@microsoft/power-apps/data';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getClient } from '@microsoft/power-apps/data';

export class CustCon_SAP_OdataService {
  private static readonly dataSourceName = 'cr_5fcustcon_5fsap_5fodata_5fe435f80a6d4c3504';

  private static readonly client = getClient(dataSourcesInfo);

  /**
   * Execute SAP OData request
   */
  public static async ExecuteSapODataRequest(api_version: string, sp: string, sv: string, body: Record<string, unknown>): Promise<IOperationResult<Record<string, unknown>>> {
    const params: { api_version: string, sp: string, sv: string, body: Record<string, unknown> } = { api_version, sp, sv, body };
    const result = await CustCon_SAP_OdataService.client.executeAsync<{ api_version: string, sp: string, sv: string, body: Record<string, unknown> }, Record<string, unknown>>(
      {
        connectorOperation: {
          tableName: CustCon_SAP_OdataService.dataSourceName,
          operationName: 'ExecuteSapODataRequest',
          parameters: params
        },
      });
    return result;
  }
}